requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
tflite-runtime==2.14.0; python_version < "3.12"
ai-edge-litert>=1.0.1; python_version >= "3.12"
